# auth-service/internal/api/http/v1/types/auth/request/register_request.go
package request

type RegisterRequest struct {
	Firstname string `json:"firstname" example:"Ivan"`
	Lastname  string `json:"lastname" example:"Ivanov"`
	Login     string `json:"login" example"ivanov@228"`
	Password  string `json:"password" example:"12345"`
}